backend/requirements.txt
azure-functions
azure-functions-durable>=1.2.4
azure-identity>=1.12.0
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=23.0.0
azure-mgmt-resource>=22.0.0
paramiko>=3.0.0
cryptography>=41.0.0
requests>=2.31.0
